+++ Toucan/Views/Components/StatusBarView.xaml.cs
@@ -10,7 +10,7 @@ public partial class StatusBarView : UserControl
         InitializeComponent();
     }
 
-    private void DefaultLanguage_Click(object sender, RoutedEventArgs e)
+    private void LanguagePanel_Click(object sender, RoutedEventArgs e)
     {
         if (sender is Button btn && btn.ContextMenu != null)
         {